Refactored code and fixed duplication on group select
