+++ src/components/icons/IconCommunity.vue
