+++ src/components/icons/IconEcosystem.vue
